refactor: convert groupMembers to data source
